feat: LOB-123: introduction of currencies in netsuite codes mapping table.

+++ src/main/java/org/cardanofoundation/lob/app/netsuite_adapter/domain/core/SearchResultTransactionItem.java
@@ -78,7 +78,7 @@ public record SearchResultTransactionItem(
 
         @JsonProperty("Currency")
         @Positive
-        Integer currency,
+        Long currency,
 
         @NotNull
         @JsonProperty("Exchange Rate")
